created a repository for respective model

+++ server/src/main/java/com/pothole_report/server/repositories/PotholeReportRepository.java
@@ -3,5 +3,5 @@ package com.pothole_report.server.repositories;
 import com.pothole_report.server.models.PotholeReport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PotholeReportRepository extends JpaRepository<PotholeReport, Integer> {
+public interface PotholeReportRepository extends JpaRepository<PotholeReport, Long> {
 }
